extract into useSearchQuery
